requirements.txt
langchain
pinecone-client
langchain-pinecone
langchain-ollama
python-dotenv
pypdf

